+++ libft/src_ft_extras/ft_abs.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_abs.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: okraus <okraus@student.42prague.com>       +#+  +:+       +#+        */
+/*   By: tlukanie <tlukanie@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/19 15:42:56 by okraus            #+#    #+#             */
-/*   Updated: 2023/03/19 15:44:01 by okraus           ###   ########.fr       */
+/*   Updated: 2024/05/09 18:44:31 by tlukanie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/libft.h"
 
-int	ft_abs(int n)
+int	ft_abs(long long n)
 {
 	if (n < 0)
 		return (-n);
